moved redux components into its own folder
